se hizo las funciones para eliminar tarea y completar en grupo.

+++ app/Http/Controllers/ControllerNota.php
@@ -78,24 +78,49 @@ class ControllerNota extends Controller
         }
     }
 
-    //Esta funcion elimina la nota, hay que probarla.
-    public function destroy(Request $request, $id)
-{
-    $user = $request->user();//Auth::user();  // Obtenemos el usuario autenticado
-    if (!$user) {
-        return response()->json(['message' => 'Usuario no autenticado'], 401);
-    }
+     //Esta funcion elimina la nota, hay que probarla.
+     public function destroy(Request $request, $id)
+     {
+         $user = $request->user();//Auth::user();  // Obtenemos el usuario autenticado
+         if (!$user) {
+             return response()->json(['message' => 'Usuario no autenticado'], 401);
+         }
 
-    // Verificamos si el usuario tiene una nota con ese ID
-    $nota = $user->notas()->find($id);
-    if (!$nota) {
-        return response()->json(['message' => 'Nota no encontrada o no perteneciente al usuario'], 404);
-    }
+         // Verificamos si el usuario tiene una nota con ese ID
+         $nota = $user->notas()->find($id);
+         if (!$nota) {
+             return response()->json(['message' => 'Nota no encontrada o no perteneciente al usuario'], 404);
+         }
 
-    // Eliminamos la nota si existe
-    $nota->delete();
-    return response()->json(['message' => 'Nota eliminada correctamente'], 200);
-}
+         // Eliminamos la nota si existe
+         $nota->delete();
+         return response()->json(['message' => 'Nota eliminada correctamente'], 200);
+     }
+
+     public function destroygrupo(Request $request, $id, $idgrupo)
+     {
+         $user = $request->user();//Auth::user();  // Obtenemos el usuario autenticado
+         if (!$user) {
+             return response()->json(['message' => 'Usuario no autenticado'], 401);
+         }
+
+         $grupo = Grupo::find($idgrupo);
+
+         if (!$grupo) {
+             return response()->json(['message' => 'La id del grupo es requerida'], 400);
+         }
+
+         // Verificamos si el usuario tiene una nota con ese ID
+         $nota = Nota::where('id', $id)->where('idgrupo', $idgrupo)->first();
+         if (!$nota) {
+             return response()->json(['message' => 'Nota no encontrada o no perteneciente al usuario'], 404);
+         }
+
+         // Eliminamos la nota si existe
+         $nota->delete();
+         return response()->json(['message' => 'Nota eliminada correctamente'], 200);
+     }
+
 
     public function show(Request $request)
     {
@@ -114,6 +139,50 @@ class ControllerNota extends Controller
     return response()->json($notas);
     }
 
+    public function updategrupo(Request $request, $id, $idgrupo)
+    {
+       // Esto busca al usuario autenticado
+        $user = $request->user();
+
+        if (!$user) {
+            return response()->json(['mensaje' => 'Usuario no autenticado'], 401);
+        }
+
+        $grupo = Grupo::find($idgrupo);
+
+        if (!$grupo) {
+            return response()->json(['message' => 'La id del grupo es requerida'], 400);
+        }
+
+        //Esto valida los datos que llegan del Frontend.
+        $datosValidados = $request->validate(
+            [
+                'estado' => 'in:Pendiente,Completada',
+                'finalizacion' => 'nullable|date_format:Y-m-d H:i:s',
+            ]
+        );
+
+        $nota = Nota::where('id', $id)->where('idgrupo', $idgrupo)->first();
+
+        //Esto actualiza los datos del usuario, si los datos se validaron correctamente.
+        if ($nota) {
+                $estado = $datosValidados['estado'] ?? $nota->estado;
+                $finalizacion = $datosValidados['finalizacion'] ?? ($estado === 'Completada' ? now() : $nota->finalizacion);
+
+            $nota->update([
+                'estado' => $estado,
+                'finalizacion' => $finalizacion,
+            ]);
+            //Esto muestra dos mensajes si se han actualizado los datos correctamente
+            //y a su vez muestra si los mismos no se pudieron actualizar.
+            return response()->json(['messaje' => 'Se ha actualizado el estado de la nota']);
+        } else {
+            return response()->json(['messaje' => 'No se ha podido actualizar el estado de la nota'], 404);
+        }
+    }
+
+
+
     public function shownotagrupo(Request $request, $id)
     {
         $user = Auth::user();
@@ -129,6 +198,7 @@ class ControllerNota extends Controller
         $notas = DB::table('notas')
                     ->select('descripcion', 'categoria', 'prioridad', 'asignacion', 'id')
                     ->where('idgrupo', '=', $id) // Solo notas que pertenecen a un grupo
+                    ->where('estado',  'Pendiente')
                     ->get();
 
         return response()->json($notas);
